requirements.txt
bcrypt==3.2.2
blinker==1.4
cffi==1.15.1
click==8.1.3
cryptography==37.0.4
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-User==1.0.2.2
Flask-WTF==1.0.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pycparser==2.21
python-dotenv==0.20.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
